--- Test_Automation_Framework_UI-develop/config/Config.properties
@@ -1,6 +1,0 @@
-timeout=10
-url=https://www.seleniumeasy.com/test/
-simpleFormUrl = https://www.seleniumeasy.com/test/basic-first-form-demo.html/
-browser=chrome
-seleniumGridURL=
-resolution=
