--- Test_Automation_Framework_UI-develop/src/main/java/selenium/utils/DriverUtil.java
@@ -1,177 +1,0 @@
-package selenium.utils;
-
-import org.apache.commons.lang3.RandomStringUtils;
-import org.openqa.selenium.*;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.ie.InternetExplorerDriver;
-import org.openqa.selenium.remote.DesiredCapabilities;
-import org.openqa.selenium.remote.RemoteWebDriver;
-
-import java.io.FileInputStream;
-import java.io.InputStream;
-import java.util.Properties;
-
-public class DriverUtil {
-    public static WebDriver driver;
-    protected static Properties properties = new Properties();
-
-    public DriverUtil(){
-
-    }
-
-    public DriverUtil(WebDriver driver) {
-        DriverUtil.driver = driver;
-    }
-
-    protected void createNewDriver() {
-        readPropertiesFile();
-        chooseBrowser();
-    }
-
-    //Choose browser from system property or properties file
-    private void chooseBrowser() {
-
-        String browser;
-
-        if (System.getProperty("browser") != null) {
-            browser = System.getProperty("browser");
-        } else {
-            browser = properties.getProperty("browser");
-        }
-
-        switch (browser){
-            case "firefox":
-                //todo
-                Log.info("Firefox browser is opening.");
-                driver.manage().window().maximize();
-                break;
-            case "chrome":
-                System.setProperty("webdriver.chrome.driver", "./src/main/resources/chromedriver.exe");
-                driver = new ChromeDriver();
-                Log.info("Chrome browser is opening.");
-                driver.manage().window().maximize();
-                break;
-            default:
-                System.setProperty("webdriver.chrome.driver", "./src/main/resources/chromedriver.exe");
-                driver = new ChromeDriver();
-                Log.info("Chrome browser is opening by default.");
-                driver.manage().window().maximize();
-                break;
-        }
-
-    }
-
-    //Locate and read data from properties file
-    private void readPropertiesFile() {
-        try {
-            InputStream input = new FileInputStream("./config/Config.properties");
-            properties.load(input);
-        } catch (Exception e) {
-            Log.error("Config.properties file not found!");
-            e.printStackTrace();
-        }
-    }
-
-    protected boolean clickToElement(WebElement element){
-        try {
-            Log.info("Clicking to the following element. ID="+element.getAttribute("id")+" , CLASS="+element.getAttribute("class")+" TEXT="+element.getText());
-            element.click();
-        }catch (NoSuchElementException e){
-            Log.error("Could not find the requested element.");
-            e.printStackTrace();
-            return false;
-        }
-        return true;
-    }
-
-    /** clearField method is clearing a textbox field by pushing the backspace button
-     * until the field is empty
-     * Params:
-     * WebElement textBox: unique ID or path to the textbox
-     */
-    protected boolean clearField(WebElement textBox){
-        try {
-            Log.info("Clearing data from the field with backspace. ID=" + textBox.getAttribute("id") + " , CLASS=" + textBox.getAttribute("class")+" NAME=" + textBox.getAttribute("name"));
-            String inputText = textBox.getAttribute("value");
-            for(int i = 0; i < inputText.length(); i++){
-                textBox.sendKeys(Keys.BACK_SPACE);
-            }
-        }catch (NoSuchElementException e){
-            Log.error("Could not find the requested textbox.");
-            e.printStackTrace();
-            return false;
-        }
-        return true;
-    }
-
-    /** writeRandomAlphabeticString method is writing a random alphabetic string into a textbox
-     * Number of characters can be specified by the user.
-     * Params:
-     * WebElement textBox: unique ID or path to the textbox
-     * int length: number of characters to be entered
-     */
-    public boolean writeRandomAlphabeticString(WebElement textBox, int length){
-        try {
-            Log.info("Writing alphabetic string into the following textbox. ID=" + textBox.getAttribute("id") + " , CLASS=" + textBox.getAttribute("class")  + " , NAME=" + textBox.getAttribute("name"));
-            textBox.sendKeys(RandomStringUtils.randomAlphabetic(length));
-        }
-        catch (NoSuchElementException e){
-            Log.error("Could not find the requested textbox.");
-            e.printStackTrace();
-            return false;
-        }
-        return true;
-    }
-
-    /** writeRandomNumericString method is writing a random numeric string into a textbox
-     * Number of characters can be specified by the user.
-     * Params:
-     * WebElement textBox: unique ID or path to the textbox
-     * int length: number of characters to be entered
-     */
-    public boolean writeRandomNumericString(WebElement textBox, int length){
-        try {
-            Log.info("Writing numeric string into the following textbox. ID=" + textBox.getAttribute("id") + " , CLASS=" + textBox.getAttribute("class")  + " , NAME=" + textBox.getAttribute("name"));
-            textBox.sendKeys(RandomStringUtils.randomNumeric(length));
-            }
-        catch (NoSuchElementException e){
-            Log.error("Could not find the requested textbox.");
-            e.printStackTrace();
-            return false;
-        }
-        return true;
-    }
-
-    /** writeRandomAlphanumericString method is writing a random alphanumberic string into a textbox
-     * Number of characters can be specified by the user.
-     * Params:
-     * WebElement textBox: unique ID or path to the textbox
-     * int length: number of characters to be entered
-     */
-    public boolean writeRandomAlphanumericString(WebElement textBox, int length){
-        try {
-            Log.info("Writing alphanumeric string into the following textbox. ID=" + textBox.getAttribute("id") + " , CLASS=" + textBox.getAttribute("class")  + " , NAME=" + textBox.getAttribute("name"));
-            textBox.sendKeys(RandomStringUtils.randomAlphanumeric(length));
-        }
-        catch (NoSuchElementException e){
-            Log.error("Could not find the requested textbox.");
-            e.printStackTrace();
-            return false;
-        }
-        return true;
-    }
-
-    public boolean clickButton(WebElement button){
-        try {
-            Log.info("Writing alphanumeric string into the following textbox. ID=" + button.getAttribute("id") + " , CLASS=" + button.getAttribute("class")  + " , NAME=" + button.getAttribute("name"));
-            button.click();
-        }
-        catch (NoSuchElementException e){
-            Log.error("Could not find the requested textbox.");
-            e.printStackTrace();
-            return false;
-        }
-        return true;
-    }
-}
