+++ config/Config.properties
@@ -3,3 +3,4 @@ url=https://www.seleniumeasy.com/test/
 browser=chrome
 seleniumGridURL=
 resolution=
+numberOfCharacters = 20
