+++ src/main/java/selenium/pageObjects/MainPageObjects.java
@@ -1,10 +1,10 @@
 package selenium.pageObjects;
 
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.PageFactory;
-import selenium.utils.DriverUtil;
+        import org.openqa.selenium.WebDriver;
+        import org.openqa.selenium.WebElement;
+        import org.openqa.selenium.support.FindBy;
+        import org.openqa.selenium.support.PageFactory;
+        import selenium.utils.DriverUtil;
 
 public class MainPageObjects extends DriverUtil {
     public MainPageObjects (WebDriver driver){
@@ -18,4 +18,12 @@ public class MainPageObjects extends DriverUtil {
     @FindBy(xpath = "//a[text()='Simple Form Demo']")
     protected WebElement simpleFormDemoSubmenuItem;
 
+    @FindBy(xpath = "//ul[@class='nav navbar-nav']//a[contains(text(),'JQuery Select dropdown')]")
+    protected WebElement jQueryDropDownSubmenuItem;
+
+    @FindBy(xpath = "//ul[@class='nav navbar-nav navbar-right']/li[4]/a[1]")
+    protected WebElement othersMenuItem;
+
+    @FindBy(xpath = "//a[text()='Drag and Drop']")
+    protected WebElement dragAndDropSubmenuItem;
 }
