+++ src/main/java/selenium/pages/MainPage.java
@@ -18,11 +18,13 @@ public class MainPage extends MainPageObjects {
             case "input_forms":
                 Assert.assertTrue("Could not click to "+menuName, clickToElement(this.inputFormsMenuItem));
                 return true;
+            case "others":
+                Assert.assertTrue("Could not click to "+menuName, clickToElement(this.othersMenuItem));
+                return true;
             default:
                 Assert.fail("Could not found the requested item: "+menuName);
                 return false;
         }
-
     }
 
     public boolean clickToSubmenuItem(String submenuName){
@@ -31,6 +33,12 @@ public class MainPage extends MainPageObjects {
         switch (submenuName){
             case "simple_form_demo":
                 Assert.assertTrue("Could not click to "+submenuName, clickToElement(this.simpleFormDemoSubmenuItem));
+                return true;
+            case "jquery_dropdown":
+                Assert.assertTrue("Could not click to "+submenuName, clickToElement(this.jQueryDropDownSubmenuItem));
+                return true;
+            case "drag_and_drop":
+                Assert.assertTrue("Could not click to "+submenuName, clickToElement(this.dragAndDropSubmenuItem));
                 return true;
             default:
                 Assert.fail("Could not found the requested item: "+submenuName);
