+++ src/test/resources/features/delete.feature
@@ -8,14 +8,14 @@ Feature: Delete text from field feature
     And Write 20 random alphabetic characters in textbox 'Message_form'
     And Delete text from 'Message_form' with backspace button
     And Click 'Show_message' button
-    Then Message is displayed in 'Your_message' field
+    Then 'Displayed message' section is empty
 
   @Sample
-  Scenario: Delete text from field with backspace button
+  Scenario: Clear text from field
     Given Open site
     When Click to 'Input_Forms' menu item
     And Click to 'Simple_Form_Demo' submenu item
     And Write 20 random alphabetic characters in textbox 'Message_form'
     And Clear data from 'Message_form' textbox
     And Click 'Show_message' button
-    Then Message is displayed in 'Your_message' field
+    Then 'Displayed message' section is empty
